Package restructuring
* With some additional minor test running tweaks
