+++ .travis.yml
@@ -1,5 +1,8 @@
+sudo: false
 language: python
-install: pip install tox
+python:
+  - "2.7"
+install: pip install tox-travis
 script: tox
 before_install: pip install codecov
 after_success: codecov
