+++ nagios_plugin/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+__all__ = ['SlaveStatusCheck']
+
+from .plugin import SlaveStatusCheck
