+++ tests/nagios_plugin/test_base.py
@@ -1,0 +1,37 @@
+# -*- coding: utf-8 -*-
+
+import unittest
+import mock
+
+from nagios_plugin.base import NagiosPlugin
+
+
+@mock.patch('nagios_plugin.base.sys')
+class TestNagiosPlugin(unittest.TestCase):
+
+    def setUp(self):
+        self.plugin_inst = NagiosPlugin(None, None)
+        self.msg = "test"
+
+    def test_run_check(self, _):
+        self.assertRaises(NotImplementedError, self.plugin_inst.run_check)
+
+    def test_ok_state(self, mock_sys):
+        self.plugin_inst.ok_state(self.msg)
+        mock_sys.exit.assert_called_once_with(0)
+        mock_sys.stdout.write.called_once_with("OK - {0}".format(self.msg))
+
+    def test_warning_state(self, mock_sys):
+        self.plugin_inst.warning_state(self.msg)
+        mock_sys.exit.assert_called_once_with(1)
+        mock_sys.stdout.write.called_once_with("WARNING - {0}".format(self.msg))
+
+    def test_critical_state(self, mock_sys):
+        self.plugin_inst.critical_state(self.msg)
+        mock_sys.exit.assert_called_once_with(2)
+        mock_sys.stdout.write.called_once_with("CRITICAL - {0}".format(self.msg))
+
+    def test_unknown_state(self, mock_sys):
+        self.plugin_inst.unknown_state(self.msg)
+        mock_sys.exit.assert_called_once_with(3)
+        mock_sys.stdout.write.called_once_with("UNKNOWN - {0}".format(self.msg))
