+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py26,py27
+envlist = py27
 
 [testenv]
 deps =
@@ -8,4 +8,4 @@ deps =
     mock
     -rrequirements.txt
 
-commands = py.test --cov=check_mariadb_slaves tests.py
+commands = pytest --cov=nagios_plugin
